% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjmg_baixar_cjsg.R
\name{tjmg_baixar_cjsg}
\alias{tjmg_baixar_cjsg}
\title{Baixar pesquisa jurisprudencial do TJMG}
\usage{
tjmg_baixar_cjsg(
  busca = NULL,
  pesquisa_por = "acordao",
  dt_pub_inicial = "",
  dt_pub_final = "",
  diretorio = "."
)
}
\arguments{
\item{busca}{Busca livre}

\item{pesquisa_por}{Por ora somente acórdão. Logo incluo monocráticas.}

\item{dt_pub_inicial}{Data inicial da publicação no formato dd/mm/yyyy}

\item{dt_pub_final}{Data final da publicação no formato dd/mm/yyy}

\item{diretorio}{Diretório onde salvar os htmls}
}
\value{
htmls
}
\description{
Baixar pesquisa jurisprudencial do TJMG
}
\details{
Se o número de decisões for muito grande, está função retornará o mesmo
aviso da página de busca do TJMG, de que se deve refinar melhor a busca.
}
