% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjmg_ler_cjsg.R
\name{tjmg_ler_cjsg}
\alias{tjmg_ler_cjsg}
\title{Lê dados da consulta jurisprudencial baixados com tjmg_baixar_cjsg}
\usage{
tjmg_ler_cjsg(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Informar arquivos}

\item{diretorio}{Se não informar arquivos, informar diretório}
}
\value{
tibble
}
\description{
Lê dados da consulta jurisprudencial baixados com tjmg_baixar_cjsg
}
